# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Ride-With-Me",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/messages/group/store": {
            "post": {
                "tags": [
                    "Group Chat"
                ],
                "summary": "Create a new group",
                "description": "Creates a new group and initializes a conversation for the group.",
                "operationId": "storeGroup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "route_id"
                                ],
                                "properties": {
                                    "route_id": {
                                        "description": "ID of the route",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Group created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Group created successfully"
                                        },
                                        "group_details": {
                                            "$ref": "#/components/schemas/Group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/members/get/{group}": {
            "get": {
                "tags": [
                    "Group Chat"
                ],
                "summary": "Retrieve all group members",
                "description": "Retrieves all members of the specified group.",
                "operationId": "retrieveAllGroupMembers",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "ID of the group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group members retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "owner": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "members": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User is not a member of the group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not a member of this group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Group not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/messages/group/send/{group}": {
            "post": {
                "tags": [
                    "Group Chat"
                ],
                "summary": "Send a message to a group",
                "description": "Sends a message to the specified group.",
                "operationId": "sendMessageToGroup",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "ID of the group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message",
                                    "type"
                                ],
                                "properties": {
                                    "message": {
                                        "description": "Message content",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of the message",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User is not a member of the group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not a member of this group"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Message": {
                "required": [
                    "content",
                    "type",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "content": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Group": {
                "title": "Group",
                "description": "Group model",
                "required": [
                    "id",
                    "route_id",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Primary key of the group",
                        "type": "string",
                        "format": "uuid"
                    },
                    "route_id": {
                        "description": "ID of the associated route",
                        "type": "string",
                        "format": "uuid"
                    },
                    "group_picture": {
                        "description": "URL of the group picture",
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status of the group",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email",
                    "password",
                    "role"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string",
                        "format": "uuid"
                    },
                    "first_name": {
                        "description": "First name of the user",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name of the user",
                        "type": "string"
                    },
                    "profile_picture": {
                        "description": "URL of the user's profile picture",
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Password of the user",
                        "type": "string"
                    },
                    "role": {
                        "description": "Role of the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}